Cosmetic: Rename heartbeat/views.py to heartbeat/api_views.py
This is more inline with our file naming scheme.
